Add Functionality to Create Asset Holders
This commit introduces the functionality to create asset holders in the Stellar network. It includes the creation of new accounts, establishing trustlines for the custom asset, and distributing the asset to these new accounts. The number of holders and the balance for each holder can be specified. This functionality is encapsulated in the createHolders function