Add Input for clawback amount and checkbox for full clawback
Add an input field to specify the amount of assets to clawback. Also, add a checkbox to enable the option of clawing back all assets. When this checkbox is selected, the input field for the clawback amount is disabled, and the clawback operation will be performed for the entire balance of the specified asset in the account.