Add Cypress test for Clawback Asset Page
Add Cypress test that simulates the clawback operation for an account on the Clawback Asset Page. The test checks the functionality of the clawback operation, including the validation of input fields and the display of success and error messages.
Network requests to Horizon Testnet are mocked using Cypress fixtures. Which also are included on this commit