Feature/recipe clawback  (#10)
* Add Clawback Feature

This commit introduces the clawback feature. The page includes input fields for the asset code, issuer secret key, and clawback account, allowing users to perform a clawback operation. And a status display area to provide feedback on the operation.

* Add Input for clawback amount and checkbox for full clawback

Add an input field to specify the amount of assets to clawback. Also, add a checkbox to enable the option of clawing back all assets. When this checkbox is selected, the input field for the clawback amount is disabled, and the clawback operation will be performed for the entire balance of the specified asset in the account.

* Add validation for clawback amount exceeding balance

This commit adds a validation check in the performClawback function to ensure that the amount specified for clawback does not exceed the available asset balance in the account. If the clawback amount is greater than the available balance, an error is thrown and the transaction is not submitted.

* Add transaction details link

Added a TransactionInfo component that displays a link to the Stellar Expert Explorer for viewing transaction details upon successful clawback operation.

* Add basic Cypress tests for Clawback Asset page

Added initial Cypress tests for the Clawback Asset page. These tests check for the existence of form elements and simulate a form submission with incorrect data. More complex tests will be added.

Add Cypress test that simulates the clawback operation for an account on the Clawback Asset Page. The test checks the functionality of the clawback operation, including the validation of input fields and the display of success and error messages.
Network requests to Horizon Testnet are mocked using Cypress fixtures. Which also are included on this commit

* Add visibility checks to Asset Page

Add assertions to check the visibility of the input fields before interacting with them.

* Refactor Cypress tests for Clawback Asset Page and add new test

This commit refactors the Cypress tests for the Clawback Asset Page to use variables instead of repeated string literals.
In addition, a new test has been added to check full clawback operation

* Add 'required' attribute to Input component

Now, with the 'required' attribute, the Input component prevents empty field submissions, reducing unnecessary requests.

* Wrap Inputs in Form and Add Required Attribute

In this commit, we have wrapped the Input components inside a form element. This allows us to use the 'required' attribute on the inputs, which prevents the form from being submitted when these inputs are empty.

* Add Cypress Test Case for Transaction Failure

In this commit, we have added a new test case in Cypress to verify that an error message is displayed when a transaction fails. We have also added a fixture for a failed transaction. This fixture is used to mock the network request for the transaction and return a simulated error response.

* Remove unnecessary statement and import

In this commit, an unnecessary @html statement and an unused import of 'Transaction' from 'stellar-sdk' were removed from the Svelte component.

The @html statement was not contributing to the functionality, and the 'Transaction' import was not being used anywhere in the code.

This do not affect the functionality of the component.

In this commit, the Cypress test was refactored to use the baseUrl from the Cypress configuration instead of hardcoding the URL.