Refactor in +page.svelte of clawback
In this commit, TypeScript support was added to the +page.svelte file of the clawback feature by adding the `lang="ts"` attribute to the `script` tag. Additionally, several variable declarations were refactored to include TypeScript type annotations. An id was also added to the Card component for better element targeting.