Refactor input value handling in Input component
Updated the Input component to handle alphanumeric input restriction internally. Introduced a new prop 'allowOnlyAlphanumeric' that, when set to true, restricts input to only alphanumeric characters. When set to false, it allows any character. This prop defaults to true, enabling alphanumeric restriction by default. This change simplifies the usage of the Input component by eliminating the need to handle this logic in parent components.