Add Customizable Sponsorship Card
Added a new Card component to the +page.svelte file to allow for customizable sponsorship operations. This Card will contain the input fields and controls necessary for users to specify the operation they want to sponsor. The logic to handle these custom operations will be implemented in subsequent commits. This change is part of an effort to make the sponsorship process more flexible and adaptable to various use cases.